+++ programs/cp-amm/src/const_pda.rs
@@ -22,6 +22,18 @@ pub mod pool_authority {
     pub const BUMP: u8 = POOL_AUTHORITY_AND_BUMP.1;
 }
 
+pub mod protocol_fee_authority {
+    use super::*;
+
+    const PROTOCOL_FEE_AUTHORITY_AND_BUMP: ([u8; 32], u8) = ed25519::derive_program_address(
+        &[crate::constants::seeds::PROTOCOL_FEE_AUTHORITY_PREFIX],
+        &crate::constants::protocol_fee_program::ID.to_bytes(),
+    );
+
+    pub const ID: Pubkey = Pubkey::new_from_array(PROTOCOL_FEE_AUTHORITY_AND_BUMP.0);
+    pub const BUMP: u8 = PROTOCOL_FEE_AUTHORITY_AND_BUMP.1;
+}
+
 // Potential optimization on event authority too since anchor internally do Pubkey::find_program_address during runtime.
 
 #[cfg(test)]