+++ CHANGELOG.md
@@ -9,13 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- New endpoint `claim_protocol_fee_unchecked`, that is similar to `claim_protocol_fee` but allows an operator with the new `ClaimProtocolFeeUnchecked` permission to withdraw from a token_vault to any token account
-
 ### Changed
+
+- `claim_protocol_fee` now requires `protocol_fee_authority` as signer instead of an operator. Only one of pool token can be claimed per instruction call.
 
 ### Deprecated
 
 ### Removed
+
+- Removed `zap_protocol_fee` endpoint
 
 ### Fixed